Fix CI build issues by refactoring ChannelBase.
